Added test cases for a scanned page (24bit image).

+++ src/algorithms/bgrtogray.cpp
@@ -7,12 +7,12 @@ void BgrToGray::apply24bit(std::shared_ptr<Image> &image)
     auto numPixelsRgb = pixelData->size();
     auto newPixelData = std::shared_ptr<ByteArray>(new ByteArray);
 
-    for (int i = 0; i < numPixelsRgb; i = i + 3) {
+    for (uint64_t i = 0; i < numPixelsRgb; i = i + 3) {
         auto b = pixelData->at(i);
         auto g = pixelData->at(i + 1);
         auto r = pixelData->at(i + 2);
 
-        auto gray =  round(0.299 * r + 0.587 * g + 0.114 * b);
+        auto gray =  static_cast<uint8_t>(round(0.299 * r + 0.587 * g + 0.114 * b));
         newPixelData->push_back(gray);
     }
